code/requirements/base.txt
amqp==2.5.0
appdirs==1.4.3
attrs==19.1.0
autopep8==1.4.4
Babel==2.7.0
billiard==3.6.0.0
bitcoin==1.1.42
cached-property==1.5.1
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==2.2.3
django-cleanup==4.0.0
django-cors-headers==3.0.2
django-filter==2.1.0
django-rest-passwordreset==1.0.0
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
drf-yasg==1.16.0
flower==0.9.3
future==0.17.1
idna==2.8
inflection==0.3.1
isodate==0.6.0
itypes==1.1.0
Jinja2==2.10.1
kombu==4.6.3
lxml==4.3.4
MarkupSafe==1.1.1
Pillow==6.1.0
psycopg2==2.8.3
pycodestyle==2.5.0
pycoin==0.90.20190630
PyJWT==1.7.1
python-dateutil==2.8.0
python-http-client==3.1.0
pytz==2019.1
redis==3.3.7
requests==2.22.0
requests-toolbelt==0.9.1
ruamel.yaml==0.15.99
sendgrid==6.0.5
sib-api-v3-sdk==6.0.2
six==1.12.0
sqlparse==0.3.0
tornado==5.1.1
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
zeep==3.4.0
